server/tests: consolidate Tax ID validation tests
Existing tests were placed under `tests/checkout`, which was clearly not appropriate.
