+++ server/tests/kit/test_tax.py
@@ -8,6 +8,7 @@ from polar.kit.address import Address, CountryAlpha2
 from polar.kit.tax import (
     InvalidTaxID,
     TaxCode,
+    TaxID,
     TaxIDFormat,
     calculate_tax,
     validate_tax_id,
@@ -85,39 +86,53 @@ class TestCalculateTaxRateLimit:
                 )
 
 
-class TestValidateTaxID:
-    def test_valid_uae_trn(self) -> None:
-        # 15-digit UAE TRN
-        result = validate_tax_id("104479084600003", "AE")
-        assert result == ("104479084600003", TaxIDFormat.ae_trn)
+@pytest.mark.parametrize(
+    ("number", "country", "expected"),
+    [
+        (
+            "GB980780684",
+            "GB",
+            (
+                "980780684",
+                TaxIDFormat.gb_vat,
+            ),
+        ),
+        ("FR61954506077", "FR", ("FR61954506077", TaxIDFormat.eu_vat)),
+        (
+            "91-1144442",
+            "US",
+            ("911144442", TaxIDFormat.us_ein),
+        ),
+        ("234567899RT0001", "CA", ("234567899RT0001", TaxIDFormat.ca_gst_hst)),
+        ("234567899 RT0001", "CA", ("234567899RT0001", TaxIDFormat.ca_gst_hst)),
+        ("234567899", "CA", ("234567899", TaxIDFormat.ca_bn)),
+        ("12.531.909-2", "CL", ("125319092", TaxIDFormat.cl_tin)),
+        ("12531909-2", "CL", ("125319092", TaxIDFormat.cl_tin)),
+        ("4540536920", "TR", ("4540536920", TaxIDFormat.tr_tin)),
+        ("27AAPFU0939F1ZV", "IN", ("27AAPFU0939F1ZV", TaxIDFormat.in_gst)),
+        ("0100233488", "VN", ("0100233488", TaxIDFormat.vn_tin)),
+        ("104479084600003", "AE", ("104479084600003", TaxIDFormat.ae_trn)),
+        ("104 479 084 600 003", "AE", ("104479084600003", TaxIDFormat.ae_trn)),
+        ("104-479-084-600-003", "AE", ("104479084600003", TaxIDFormat.ae_trn)),
+    ],
+)
+def test_validate_tax_id_valid(number: str, country: str, expected: TaxID) -> None:
+    validated_tax_id = validate_tax_id(number, country)
+    assert validated_tax_id == expected
 
-    def test_valid_uae_trn_with_separators(self) -> None:
-        # UAE TRN with spaces and dashes should be normalized
-        result = validate_tax_id("104 479 084 600 003", "AE")
-        assert result == ("104479084600003", TaxIDFormat.ae_trn)
 
-        result = validate_tax_id("104-479-084-600-003", "AE")
-        assert result == ("104479084600003", TaxIDFormat.ae_trn)
-
-    def test_invalid_uae_trn_wrong_length(self) -> None:
-        # Too short
-        with pytest.raises(InvalidTaxID):
-            validate_tax_id("12345678901234", "AE")
-
-        # Too long
-        with pytest.raises(InvalidTaxID):
-            validate_tax_id("1234567890123456", "AE")
-
-    def test_invalid_uae_trn_non_numeric(self) -> None:
-        with pytest.raises(InvalidTaxID):
-            validate_tax_id("10447908460000A", "AE")
-
-    def test_valid_eu_vat(self) -> None:
-        # French VAT number
-        result = validate_tax_id("FR61954506077", "FR")
-        assert result == ("FR61954506077", TaxIDFormat.eu_vat)
-
-    def test_invalid_country(self) -> None:
-        # Country not in COUNTRY_TAX_ID_MAP
-        with pytest.raises(InvalidTaxID):
-            validate_tax_id("123456789", "XX")
+@pytest.mark.parametrize(
+    ("number", "country"),
+    [
+        ("123", "FR"),
+        ("FR11111111111", "FR"),
+        ("GB980780684", "FR"),
+        ("GB980780684", "foo"),
+        ("10447908460000A", "AE"),
+        ("10447908460000", "AE"),
+        ("1044790846000000", "AE"),
+    ],
+)
+def test_validate_tax_id_invalid(number: str, country: str) -> None:
+    with pytest.raises(InvalidTaxID):
+        validate_tax_id(number, country)
